+++ main/main.h
@@ -48,6 +48,10 @@ public:
 	static int test_entrypoint(int argc, char *argv[], bool &tests_need_run);
 	static Error setup(const char *execpath, int argc, char *argv[], bool p_second_phase = true);
 	static Error setup2(Thread::ID p_main_tid_override = 0);
+#ifdef TESTS_ENABLED
+	static Error test_setup();
+	static void test_cleanup();
+#endif
 	static bool start();
 
 	static bool iteration();
@@ -58,7 +62,7 @@ public:
 	static void cleanup();
 };
 
-// Test main override is for the testing behaviour
+// Test main override is for the testing behaviour.
 #define TEST_MAIN_OVERRIDE                                         \
 	bool run_test = false;                                         \
 	int return_code = Main::test_entrypoint(argc, argv, run_test); \
