+++ modules/mobile_vr/register_types.cpp
@@ -35,9 +35,11 @@
 void register_mobile_vr_types() {
 	ClassDB::register_class<MobileVRInterface>();
 
-	Ref<MobileVRInterface> mobile_vr;
-	mobile_vr.instance();
-	XRServer::get_singleton()->add_interface(mobile_vr);
+	if (XRServer::get_singleton()) {
+		Ref<MobileVRInterface> mobile_vr;
+		mobile_vr.instance();
+		XRServer::get_singleton()->add_interface(mobile_vr);
+	}
 }
 
 void unregister_mobile_vr_types() {
